Update README.md
The Erosion Control Project in Rigachikun Community aims to address the immediate challenges caused by erosion and riverbank, impacting local farmers' land and the accessibility of the only road in the area. The project is currently at the pilot stage, testing the feasibility and effectiveness of erosion control measures, including the construction of a retention wall and a temporary bridge, to mitigate the effects of climate change and protect the community infrastructure.