Update README.md
Implementation Plan:
The project involves the installation of a retention wall along the riverbank to control water flow and prevent erosion, as well as the construction of a temporary bridge to provide access and manage water movement in the area. Community members and stakeholders are actively engaged in the project, providing insights, feedback, and support throughout the implementation process 

          Monitoring and Evaluation:
A monitoring and evaluation plan has been established to track the performance of the erosion control measures, gather feedback from the community, and assess the impact of the  project on protecting farmland, ensuring road access, and strengthening community resilience against erosion.

        Next Steps:
Based on the findings and outcomes of the pilot project, further scaling and expansion of the erosion control measures may be considered to address the broader erosion challenges in Rigachikun Community. Regular updates, reports, and collaboration with stakeholders will drive the project forward towards sustainable and effective solutions for erosion control in the community.